Moved exact_match_utils updates to exact_match module
